Implements the atom publishing at each form submission
